Bugfix: fix race-condition on velbus_query
There is no guarantee that the finally:-block of velbus_query() will run
before the next message will be processed. This could lead to the
message_filter() being run again, trying to .set_result() the same
future again, raising an InvalidState error.

This patch ensures that only the first reply is ever used